# contracts/prototype-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ceramic Glaze Lab — Local Research Prototype",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/reactions/calcite": {
      "post": {
        "summary": "Calcite Explorer",
        "operationId": "calcite_explorer_api_v1_reactions_calcite_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactionExplorerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "type": "object",
                  "title": "Response Calcite Explorer Api V1 Reactions Calcite Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/validation/temperature": {
      "post": {
        "summary": "Temperature Comparison",
        "operationId": "temperature_comparison_api_v1_validation_temperature_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparisonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "type": "object",
                  "title": "Response Temperature Comparison Api V1 Validation Temperature Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outcomes/assess": {
      "post": {
        "summary": "Outcome Indicators",
        "operationId": "outcome_indicators_api_v1_outcomes_assess_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutcomeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "type": "object",
                  "title": "Response Outcome Indicators Api V1 Outcomes Assess Post"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/research/archive/recipes": {
      "get": {
        "summary": "Research Archive Recipes",
        "operationId": "research_archive_recipes_api_v1_research_archive_recipes_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "title": "Response Research Archive Recipes Api V1 Research Archive Recipes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/archive/recipes/{source_id}": {
      "get": {
        "summary": "Research Archive Recipe",
        "operationId": "research_archive_recipe_api_v1_research_archive_recipes__source_id__get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "title": "Response Research Archive Recipe Api V1 Research Archive Recipes  Source Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses": {
      "post": {
        "summary": "Analyze",
        "operationId": "analyze_api_v1_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisReport"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisReport": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "prototype-report/1",
            "title": "Schema Version",
            "default": "prototype-report/1"
          },
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "request": {
            "$ref": "#/components/schemas/AnalysisRequest"
          },
          "chemistry": {
            "$ref": "#/components/schemas/ChemistryResult"
          },
          "process": {
            "$ref": "#/components/schemas/ProcessReport"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/Material"
            },
            "type": "array",
            "title": "Materials"
          },
          "notices": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notices"
          },
          "persistence": {
            "type": "string",
            "const": "NOT_STORED_EXPORT_TO_KEEP",
            "title": "Persistence",
            "default": "NOT_STORED_EXPORT_TO_KEEP"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "report_id",
          "request",
          "chemistry",
          "process",
          "materials",
          "notices"
        ],
        "title": "AnalysisReport"
      },
      "AnalysisRequest": {
        "properties": {
          "recipe_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Recipe Name"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/Ingredient"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Ingredients"
          },
          "base_mass_g": {
            "type": "number",
            "maximum": 1000000.0,
            "minimum": 1e-06,
            "title": "Base Mass G",
            "default": 100
          },
          "context": {
            "$ref": "#/components/schemas/FiringContext"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "recipe_name",
          "ingredients"
        ],
        "title": "AnalysisRequest"
      },
...
      "ComparisonRequest": {
        "properties": {
          "prediction": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Prediction"
          },
          "observation": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Observation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "prediction",
          "observation"
        ],
        "title": "ComparisonRequest"
      },
      "ErrorItem": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "path": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Path"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "severity": {
            "type": "string",
            "const": "ERROR",
            "title": "Severity",
            "default": "ERROR"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "path",
          "message"
        ],
        "title": "ErrorItem"
      },
      "ErrorResponse": {
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "errors"
        ],
        "title": "ErrorResponse"
      },
      "FiringContext": {
        "properties": {
          "cone": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "06",
              "04",
              "6",
              "8",
              "10"
            ],
            "title": "Cone",
            "default": "UNKNOWN"
          },
          "temperature_c": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1800.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature C"
          },
          "atmosphere": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "OXIDATION",
              "REDUCTION",
              "OTHER"
            ],
            "title": "Atmosphere",
            "default": "UNKNOWN"
          },
          "clay_body": {
            "type": "string",
            "maxLength": 160,
            "title": "Clay Body",
            "default": ""
          },
          "body_window": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportedWindow"
              },
              {
                "type": "null"
              }
            ]
          },
          "glaze_window": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportedWindow"
              },
              {
                "type": "null"
              }
            ]
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FiringSchedule"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "FiringContext"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Ingredient": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Analysis Id"
          },
          "amount": {
            "type": "number",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Amount"
          },
          "role": {
            "type": "string",
            "enum": [
              "BASE",
              "ADDITION"
            ],
            "title": "Role",
            "default": "BASE"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "analysis_id",
          "amount"
        ],
        "title": "Ingredient"
      },
      "JsonValue": {},
...
      "OutcomeRequest": {
        "properties": {
          "fit": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fit"
          },
          "flow": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flow"
          },
          "wetting": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wetting"
          },
          "porosity": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Porosity"
          },
          "gloss": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gloss"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "OutcomeRequest"
      },
...
      "ProcessReport": {
        "properties": {
          "engine_version": {
            "type": "string",
            "title": "Engine Version"
          },
          "input_hash": {
            "type": "string",
            "title": "Input Hash"
          },
          "input_snapshot": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Input Snapshot"
          },
          "status": {
            "type": "string",
            "const": "PARTIAL",
            "title": "Status"
          },
          "checks_evidence_kind": {
            "type": "string",
            "const": "CALCULATED",
            "title": "Checks Evidence Kind"
          },
          "checks_method_kind": {
            "type": "string",
            "const": "DETERMINISTIC",
            "title": "Checks Method Kind"
          },
          "checks_qualifier": {
            "type": "string",
            "title": "Checks Qualifier"
          },
          "body_window": {
            "$ref": "#/components/schemas/WindowCheck"
          },
          "glaze_window": {
            "$ref": "#/components/schemas/WindowCheck"
          },
          "schedule": {
            "$ref": "#/components/schemas/ScheduleCheck"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/ProcessStage"
            },
            "type": "array",
            "title": "Stages"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "engine_version",
          "input_hash",
          "input_snapshot",
          "status",
          "checks_evidence_kind",
          "checks_method_kind",
          "checks_qualifier",
          "body_window",
          "glaze_window",
          "schedule",
          "stages",
          "warnings",
          "limitations"
        ],
        "title": "ProcessReport"
      },
...
      "ReactionExplorerRequest": {
        "properties": {
          "mass_g": {
            "type": "number",
            "maximum": 1000000.0,
            "minimum": 1e-06,
            "title": "Mass G"
          },
          "conversion": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Conversion"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mass_g",
          "conversion"
        ],
        "title": "ReactionExplorerRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}